feat(Game Logic Function): Add function to take player and computer choices. adds 1 to score and logs winner
I created 4 global variables ln 45, 46, 63,64 to store the choices of the human and computer . Then two variables to keep track of player score, and initailaized to a value of 0. A function playround was made that compared to if the human/computer choices were the same, or different and then would increment the winner's score in the global variable.