feat(Game logic for round count): Build function to play entire game to 5 rounds and change the score
Made function playGame with 3 variables, one to call playRound, and the other two to call the score values. This function will play for 5 rounds and with a loop, it sets the round count variable, and then keeps track of the rounds to be less than 5. For every loop, it will increment the round cound 1. If either score reaches 5, then the game will  reset the variables for both humanScore, and computerScore.  I also slightly changed function playRound to figure out some logic errors.